requirements.txt
# Web Framework
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
Werkzeug==2.0.1
gunicorn==20.1.0
flask-swagger-ui==3.36.0

# Data Processing
pandas==1.3.3
numpy==1.21.2
scikit-learn==1.0.0
joblib==1.1.0
faiss-cpu==1.7.2
thefuzz==0.19.0

# NLP
spacy==3.2.0
sentence-transformers==2.1.0
nltk==3.6.5

# Image Processing
Pillow==8.4.0
tensorflow==2.8.0
tensorflow-hub==0.12.0

# Database
SQLAlchemy==1.4.26

# Utilities
python-dotenv==0.19.1
requests==2.26.0
urllib3==1.26.7
tqdm==4.62.3